Bounce Rate 이상 수치 에러 해결

+++ src/main/java/com/junction/tonight/spark/dto/BaseDataFormat.java
@@ -14,7 +14,7 @@ import java.util.HashMap;
 @AllArgsConstructor
 public class BaseDataFormat {
 
-    public Integer totalNumber;
+    public Double totalNumber;
 
     public HashMap<String, Integer> areaData;
 }
